Use SH_INITIALIZE_UNINITIALIZED_LOCALS now that ANGLE has been updated https://bugs.webkit.org/show_bug.cgi?id=177929 <rdar://problem/34866210>
Reviewed by Sam Weinig.

Source/WebCore:

Test: fast/canvas/webgl/uninitialized-locals.html

* platform/graphics/ANGLEWebKitBridge.cpp:
(WebCore::ANGLEWebKitBridge::compileShaderSource):
* platform/graphics/ANGLEWebKitBridge.h:
* platform/graphics/opengl/Extensions3DOpenGLCommon.cpp:
(WebCore::Extensions3DOpenGLCommon::getTranslatedShaderSourceANGLE):

LayoutTests:

* fast/canvas/webgl/uninitialized-locals-expected.txt: Added.
* fast/canvas/webgl/uninitialized-locals.html: Added.

Canonical link: https://commits.webkit.org/196256@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225407 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/ANGLEWebKitBridge.h
@@ -73,7 +73,7 @@ public:
     const ShBuiltInResources& getResources() { return m_resources; }
     void setResources(const ShBuiltInResources&);
     
-    bool compileShaderSource(const char* shaderSource, ANGLEShaderType, String& translatedShaderSource, String& shaderValidationLog, Vector<std::pair<ANGLEShaderSymbolType, sh::ShaderVariable>>& symbols, int extraCompileOptions = 0);
+    bool compileShaderSource(const char* shaderSource, ANGLEShaderType, String& translatedShaderSource, String& shaderValidationLog, Vector<std::pair<ANGLEShaderSymbolType, sh::ShaderVariable>>& symbols, long extraCompileOptions = 0);
 
 private:
